moving az_des member_name redaction to the CbvApplicant::AzDes subclass
